requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
python-jose[cryptography]
passlib[bcrypt]
redis
pytest
pytest-asyncio
pytest-cov
httpx
python-dotenv
fakeredis
